pyproject.toml
[project]
name = "pm-ai-toolkit"
version = "0.1.0"
description = "Shared foundation for AI-powered PM tools and workflows"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "anthropic>=0.25",
    "openai>=1.0",
    "pypdf>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
